requirements.txt
numpy==1.25.0
pandas==2.0.2
Pillow==11.0.0
rembg==2.0.60
scikit_learn==1.3.2
selenium==4.27.1
tabulate==0.9.0
torch==2.5.0
torchvision==0.20.0
tqdm==4.66.5
tqdm==4.66.4
webdriver_manager==4.0.2
